Trackers
These are just some webhooks that track text using requests or BeatifulSoup(I forgot to mention that it uses gtts to read them out loud to me in text to speach but an AI helped me with that). I used it to track the headlines of news on one high school in Prague. And the 2nd one just tracks one line of text so I know when it disappears (its very useful because lets say that I really want to be the first one to know about TechDays but they dont accept registrations yet but because it tracks it every hour when the text we dont accept registrations yet disappears I will know that I need to put in my registration and that they accept them now).